+++ src/settings.ts
@@ -47,6 +47,7 @@ export interface AbstractFolderPluginSettings {
   namingConflictSeparator: '-' | 'brackets'; // Separator to use for naming conflicts
   namingConflictOrder: 'parent-first' | 'name-first'; // Order of parent and name
   defaultNewNotePath: string; // Default path for new notes
+  anonymizeDebugExport: boolean;
 }
 
 export const DEFAULT_SETTINGS: AbstractFolderPluginSettings = {
@@ -96,4 +97,5 @@ export const DEFAULT_SETTINGS: AbstractFolderPluginSettings = {
   namingConflictSeparator: '-',
   namingConflictOrder: 'parent-first',
   defaultNewNotePath: '',
+  anonymizeDebugExport: true,
 };